rename test, fix project dependencies
